Edit Todo feature added + project delete button right align css
Able to edit todo's, and it will bring up the new todo dialog but
pre-populate the fields. Once editing is done, can hit enter and
it will write over the old todo in the Map.

Also moved the delete button on the project sidebar so that its
along the right end.